Extract transform-and-validate step and ledger id
Move the Level 4 transform + data-business validation out of validate.js
into its own module, folding the previous-submission row-state read in
beside the existing-records read so both prior-state reads live together.
Extract the ledger identity a summary log's reads pivot on as a typed
WasteBalanceLedgerId, carrying the full organisation/registration/
accreditation chain rather than the accreditation slot alone.
